% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timstools.R
\name{sink_session_info}
\alias{sink_session_info}
\title{Save session information to file}
\usage{
sink_session_info(additional_filename_text = "", include_date = TRUE)
}
\arguments{
\item{additional_filename_text}{User-specified text to be included in the
saved file name.}

\item{include_date}{Boolean. Default=TRUE. If TRUE, include today's date in
the saved file name in format \%Y-\%m-\%d.}
}
\value{
Saves a text file with \code{sessionInfo()} to the current directory.
}
\description{
Save \code{sink()} session information to a text file in the working directory.
}
\examples{
sink_session_info(additional_filename_text = my_R_project,
                  include_date = FALSE)

}
